back-end/estudiantes/requirements.txt
# requirements.txt - Dependencias del proyecto
Flask==3.0.0
Flask-CORS==4.0.0
mysql-connector-python==8.2.0
python-dotenv==1.0.0
gunicorn==21.2.0
python-dateutil==2.8.2

# Dependencias para desarrollo (opcional)
pytest==7.4.3
pytest-flask==1.3.0
black==23.11.0
flake8==6.1.0